11-svc/pyproject.toml
[project]
name = "11-svc"
version = "0.1.0"
description = "svc"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "llama-index-embeddings-ollama>=0.6.0",
    "llama-index-llms-ollama>=0.6.2",
    "llama-index-node-parser-topic>=0.2.0",
    "llama-index-vector-stores-qdrant>=0.6.1",
]
